requirements.txt

# Django
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Configuration
python-decouple==3.8

# Distributed Locking
python-redis-lock==4.0.0

# Monitoring
django-prometheus==2.3.1
prometheus-client==0.19.0

# API Clients
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
ipython==8.17.2
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# CLI
tabulate==0.9.0